progress update

+++ controllers/trade_controllers.js
@@ -1,19 +1,8 @@
 const express = require("express");
 const db = require("../models");
 const router = express.Router();
+const helper = require("../public/javascript/app");
 const userid = "";
-
-// //$('document').on('click', function() {
-//   FB.getLoginStatus(function(response) {
-//     console.log("loginstatus just ran");
-//     userid = response.authResponse.userID;
-//
-//     console.log("your userID is " + userid);
-//
-//     console.log(response);
-//     console.log(response.status + "is your current login state with facebook.");
-//   });
-// });
 
 router.get("/", (req, res) => {
   res.redirect("/trades");
@@ -46,8 +35,9 @@ router.post("/trades/api", (req, res) => {
     transmission: req.body.transmission,
     description: req.body.description,
     contact: req.body.contact,
+
     zipcode: req.body.zipcode,
-    userID: userid
+    userID: req.body.userID
   }).then(() => {
     return res.redirect("/trades");
   }).catch((err) => {
@@ -56,13 +46,16 @@ router.post("/trades/api", (req, res) => {
 });
 
 router.post("/trades/:id", (req, res) => {
-  db.trades.destroy({
-    where: {
-      id: req.params.id
-    }
-  }).then((result) => {
-    return res.redirect("/trades");
-  });
+  if (this.userID === currentuserID) {
+    db.trades.destroy({
+      where: {
+        id: req.params.id
+      }
+    }).then((result) => {
+      return res.redirect("/trades");
+    });
+  } else {
+    alert("You don't own this post!")
+  };
 });
-
 module.exports = router;
